requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.6
langgraph==0.0.20
google-generativeai==0.3.2
faiss-cpu==1.7.4
pytesseract==0.3.10
pymupdf==1.23.8
pandas==2.1.4
duckduckgo-search==3.8.6
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
openpyxl==3.1.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
streamlit==1.28.1
plotly==5.17.0

